review: move quicknode and moralis logic into separate crates, pin redis:7.2.4-alpine3.19 version. fix: Execute the pipeline once after setting all commands and make upsert_address_rate_in_pipe not async, as it just mutates Pipeline.
