+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.8'
 services:
   redis:
-    image: redis:latest
+    image: redis:7.2.4-alpine3.19
     restart: always
     ports:
       - "6379:6379"
