+++ src/net/server.rs
@@ -3,19 +3,12 @@ use std::str::FromStr;
 
 use hyper::server::conn::AddrStream;
 use hyper::service::{make_service_fn, service_fn};
-use hyper::{Body, Request, Response, Server, StatusCode, Uri};
+use hyper::{Body, Request, Response, Server, StatusCode};
 
 use super::{GenericError, GenericResult};
-use crate::address_status::{AddressStatus, AddressStatusOperations};
-use crate::ctx::{AppConfig, ProxyRoute};
-use crate::db::Db;
-use crate::http::{
-    http_handler, response_by_status, HttpGetPayload, JsonRpcPayload, PayloadData, X_FORWARDED_FOR,
-};
+use crate::ctx::AppConfig;
+use crate::http::{http_handler, response_by_status, X_FORWARDED_FOR};
 use crate::log_format;
-use crate::proof_of_funding::{verify_message_and_balance, ProofOfFundingError};
-use crate::rate_limiter::RateLimitOperations;
-use crate::sign::SignOps;
 use crate::websocket::{should_upgrade_to_socket_conn, socket_handler};
 
 #[macro_export]
@@ -73,238 +66,6 @@ async fn connection_handler(
         socket_handler(cfg.clone(), req, remote_addr).await
     } else {
         http_handler(cfg, req, remote_addr).await
-    }
-}
-
-pub(crate) async fn validation_middleware(
-    cfg: &AppConfig,
-    payload: &PayloadData,
-    proxy_route: &ProxyRoute,
-    req_uri: &Uri,
-    remote_addr: &SocketAddr,
-) -> Result<(), StatusCode> {
-    match payload {
-        PayloadData::JsonRpc(payload) => {
-            validation_middleware_json_rpc(cfg, payload, proxy_route, req_uri, remote_addr).await
-        }
-        PayloadData::HttpGet(payload) => {
-            validation_middleware_http_get(cfg, payload, proxy_route, req_uri, remote_addr).await
-        }
-    }
-}
-
-pub(crate) async fn validation_middleware_json_rpc(
-    cfg: &AppConfig,
-    payload: &JsonRpcPayload,
-    proxy_route: &ProxyRoute,
-    req_uri: &Uri,
-    remote_addr: &SocketAddr,
-) -> Result<(), StatusCode> {
-    let mut db = Db::create_instance(cfg).await;
-
-    match db
-        .read_address_status(&payload.signed_message.address)
-        .await
-    {
-        AddressStatus::Trusted => Ok(()),
-        AddressStatus::Blocked => Err(StatusCode::FORBIDDEN),
-        AddressStatus::None => {
-            let signed_message_status = verify_message_and_balance(cfg, payload, proxy_route).await;
-
-            if let Err(ProofOfFundingError::InvalidSignedMessage) = signed_message_status {
-                log::warn!(
-                    "{}",
-                    log_format!(
-                        remote_addr.ip(),
-                        payload.signed_message.address,
-                        req_uri,
-                        "Request has invalid signed message, returning 401"
-                    )
-                );
-
-                return Err(StatusCode::UNAUTHORIZED);
-            };
-
-            let rate_limiter_key = format!(
-                "{}:{}",
-                payload.signed_message.coin_ticker, payload.signed_message.address
-            );
-
-            let rate_limiter = proxy_route
-                .rate_limiter
-                .as_ref()
-                .unwrap_or(&cfg.rate_limiter);
-            match db.rate_exceeded(&rate_limiter_key, rate_limiter).await {
-                Ok(false) => {}
-                _ => {
-                    log::warn!(
-                        "{}",
-                        log_format!(
-                            remote_addr.ip(),
-                            payload.signed_message.address,
-                            req_uri,
-                            "Rate exceed for {}, checking balance for {} address.",
-                            rate_limiter_key,
-                            payload.signed_message.address
-                        )
-                    );
-
-                    match verify_message_and_balance(cfg, payload, proxy_route).await {
-                        Ok(_) => {}
-                        Err(ProofOfFundingError::InsufficientBalance) => {
-                            log::warn!(
-                                "{}",
-                                log_format!(
-                                    remote_addr.ip(),
-                                    payload.signed_message.address,
-                                    req_uri,
-                                    "Wallet {} has insufficient balance for coin {}, returning 406.",
-                                    payload.signed_message.address,
-                                    payload.signed_message.coin_ticker,
-                                )
-                            );
-
-                            return Err(StatusCode::NOT_ACCEPTABLE);
-                        }
-                        e => {
-                            log::error!(
-                                "{}",
-                                log_format!(
-                                    remote_addr.ip(),
-                                    payload.signed_message.address,
-                                    req_uri,
-                                    "verify_message_and_balance failed in coin {}: {:?}",
-                                    payload.signed_message.coin_ticker,
-                                    e
-                                )
-                            );
-                            return Err(StatusCode::INTERNAL_SERVER_ERROR);
-                        }
-                    }
-                }
-            };
-
-            if db.rate_address(rate_limiter_key).await.is_err() {
-                log::error!(
-                    "{}",
-                    log_format!(
-                        remote_addr.ip(),
-                        payload.signed_message.address,
-                        req_uri,
-                        "Rate incrementing failed."
-                    )
-                );
-            };
-
-            Ok(())
-        }
-    }
-}
-
-pub(crate) async fn validation_middleware_http_get(
-    cfg: &AppConfig,
-    payload: &HttpGetPayload,
-    proxy_route: &ProxyRoute,
-    req_uri: &Uri,
-    remote_addr: &SocketAddr,
-) -> Result<(), StatusCode> {
-    let mut db = Db::create_instance(cfg).await;
-
-    match db
-        .read_address_status(&payload.signed_message.address)
-        .await
-    {
-        AddressStatus::Trusted => Ok(()),
-        AddressStatus::Blocked => Err(StatusCode::FORBIDDEN),
-        AddressStatus::None => {
-            match payload.signed_message.verify_message() {
-                Ok(true) => {}
-                Ok(false) => {
-                    log::warn!(
-                        "{}",
-                        log_format!(
-                            remote_addr.ip(),
-                            payload.signed_message.address,
-                            req_uri,
-                            "Request has invalid signed message, returning 401"
-                        )
-                    );
-
-                    return Err(StatusCode::UNAUTHORIZED);
-                }
-                Err(e) => {
-                    log::error!(
-                        "{}",
-                        log_format!(
-                            remote_addr.ip(),
-                            payload.signed_message.address,
-                            req_uri,
-                            "verify_message failed in coin {}: {}, returning 500.",
-                            payload.signed_message.coin_ticker,
-                            e
-                        )
-                    );
-                    return Err(StatusCode::INTERNAL_SERVER_ERROR);
-                }
-            }
-
-            let rate_limiter_key = format!(
-                "{}:{}",
-                payload.signed_message.coin_ticker, payload.signed_message.address
-            );
-
-            let rate_limiter = proxy_route
-                .rate_limiter
-                .as_ref()
-                .unwrap_or(&cfg.rate_limiter);
-            match db.rate_exceeded(&rate_limiter_key, rate_limiter).await {
-                Ok(false) => {}
-                Ok(true) => {
-                    log::warn!(
-                        "{}",
-                        log_format!(
-                            remote_addr.ip(),
-                            payload.signed_message.address,
-                            req_uri,
-                            "Rate exceed for {}, returning 406.",
-                            rate_limiter_key,
-                        )
-                    );
-                    return Err(StatusCode::NOT_ACCEPTABLE);
-                }
-                Err(e) => {
-                    log::error!(
-                        "{}",
-                        log_format!(
-                            remote_addr.ip(),
-                            payload.signed_message.address,
-                            req_uri,
-                            "Rate exceeded check failed in coin {}: {}, returning 500.",
-                            payload.signed_message.coin_ticker,
-                            e
-                        )
-                    );
-                    return Err(StatusCode::INTERNAL_SERVER_ERROR);
-                }
-            }
-
-            if let Err(e) = db.rate_address(rate_limiter_key).await {
-                log::error!(
-                    "{}",
-                    log_format!(
-                        remote_addr.ip(),
-                        payload.signed_message.address,
-                        req_uri,
-                        "Rate incrementing failed in coin {}: {}, returning 500.",
-                        payload.signed_message.coin_ticker,
-                        e
-                    )
-                );
-                return Err(StatusCode::INTERNAL_SERVER_ERROR);
-            };
-
-            Ok(())
-        }
     }
 }
